+++ cluster/main.tf
@@ -5,6 +5,7 @@ resource "aws_ecs_cluster" "main" {
 resource "aws_cloudwatch_log_group" "ecs_agent" {
   name              = "${var.log_group}"
   retention_in_days = "${var.log_groups_expiration_days}"
+  tags              = "${var.log_groups_tags}"
 }
 
 resource "aws_autoscaling_group" "app" {