+++ service_load_balancing/main.tf
@@ -55,4 +55,5 @@ resource "aws_cloudwatch_log_group" "app" {
 
   name              = "${var.log_groups[count.index]}"
   retention_in_days = "${var.log_groups_expiration_days}"
+  tags              = "${var.log_groups_tags}"
 }